Tidy up

+++ src/main/java/net/sf/mpxj/mpp/MPP9Reader.java
@@ -678,7 +678,7 @@ final class MPP9Reader implements MPPVariantReader
          return;
       }
 
-      Integer prev = 0;
+      Integer prev = Integer.valueOf(0);
       int value = MPPUtility.getInt(data, uniqueIDOffset);
       while (value != SUBPROJECT_LISTEND)
       {
@@ -699,7 +699,7 @@ final class MPP9Reader implements MPPVariantReader
 
             default:
             {
-               if (prev != 0)
+               if (prev.intValue()!= 0)
                {
                   // The previous value was for an external task unique task id
                   sp.addExternalTaskUniqueID(prev);
